Implemented [fetchTopIds] in the db provider, removed debug prints

+++ lib/src/resources/news_db_provider.dart
@@ -11,6 +11,9 @@ class NewsDbProvider implements Cache, Source {
   /// A connection to the database.
   Database db;
 
+  /// Flag that indicates if the database has been initialized.
+  bool get isReady => db != null;
+
   NewsDbProvider() {
     init();
   }
@@ -55,12 +58,26 @@ class NewsDbProvider implements Cache, Source {
 
   // TODO: implement this functionality.
   // Returns a list of top ids from the local database.
-  Future<List<int>> fetchTopIds() {
+  Future<List<int>> fetchTopIds() async {
+    if (!isReady) {
+      return null;
+    }
+    final maps = await db.query(
+      "TopIds",
+      where: null,
+    );
+    if (maps.length > 0) {
+      return mapDbMapsToListOfIds(maps);
+    }
     return null;
   }
 
   /// Returns an item from a particular id.
   Future<ItemModel> fetchItem(int id) async {
+    if (!isReady) {
+      return null;
+    }
+
     final maps = await db.query(
       "Items",
       where: "id = ?",
@@ -75,14 +92,53 @@ class NewsDbProvider implements Cache, Source {
   }
 
   /// Inserts an item into the local database.
-  Future<int> addItem(ItemModel item) {
+  Future<int> addItem(ItemModel item) async {
+    // We dont want to attempt to insert this item if the db is not ready or if
+    // the item provided is null.
+    if (!isReady || item == null) {
+      return 0;
+    }
     return db.insert("Items", item.toDbMap());
+  }
+
+  Future<void> addTopIds(List<int> ids) async {
+    // We dont want to attempt to add the ids if the db is not ready or if the
+    // data provided is null.
+    if (!isReady || ids == null) {
+      return null;
+    }
+    final batch = db.batch();
+    ids.forEach(
+      (int id) {
+        batch.insert(
+          "TopIds",
+          {
+            "id": id,
+          },
+        );
+      },
+    );
+    await batch.commit();
   }
 
   /// Deletes of the data inside the tables.
   Future<void> clear() async {
+    if (!isReady) {
+      return;
+    }
     await db.delete('Items');
     await db.delete('TopIds');
+  }
+
+  List<int> mapDbMapsToListOfIds(List<Map<String, dynamic>> maps) {
+    return maps
+        .map(
+          (map) {
+            return map["id"];
+          },
+        )
+        .toList()
+        .cast<int>();
   }
 }
